Server: YAML Parsing for Configs Architecture
Architecture for when implementing a new config:
- Create a new XCfg struct
- Create the .yaml in the server/configs
- Extend the Config<T>::parse_file() for the new XCfg
- Add as a member variable to where the cfg wants to be attached
- On its contruction the parse_file will be called

+++ server/src/game/model/core/Instance.cpp
@@ -1,8 +1,8 @@
 #include "Instance.h"
 
-Instance::Instance(size_t id) : id(id), is_active(true) {}
+Instance::Instance(int id) : id(id), is_active(true) {}
 
-size_t Instance::Id() const {
+int Instance::Id() const {
     return id;
 }
