+++ server/src/game/systems/updatables/UpdatablesSystem.cpp
@@ -4,7 +4,7 @@ UpdatablesSystem::UpdatablesSystem(int rate): rate(rate)  {}
 
 void UpdatablesSystem::update(
     const int it,
-    const std::unordered_map<size_t, std::shared_ptr<Worm>> &worms,
+    const std::unordered_map<int, std::shared_ptr<Worm>> &worms,
     const std::vector<std::shared_ptr<Projectile>> &projectiles
 ) const {
     for (const auto& [_, worm]: worms) {
