+++ server/src/game/systems/updatables/UpdatablesSystem.h
@@ -11,7 +11,7 @@ public:
     UpdatablesSystem(int rate);
     void update(
         int it,
-        const std::unordered_map<size_t, std::shared_ptr<Worm>> &worms,
+        const std::unordered_map<int, std::shared_ptr<Worm>> &worms,
         const std::vector<std::shared_ptr<Projectile>> &projectiles
         ) const;
 };
